thêm nút từ cơ sở dữ liệu sang Animal sửa model Animal khóa ngoài thêm View Index cho Plant !!! khi bấm vào View Animal từ AnimalFacility không hiện gì

+++ quan_ly_tai_nguyen_rung/Models/section4/AnimalFacility.cs
@@ -41,7 +41,7 @@ namespace quan_ly_tai_nguyen_rung.Models.section4
 
         [Required]
         [Column("ACREAGE")]
-        public float Acreage { get; set; } // DIỆN TÍCH 
+        public double Acreage { get; set; } // DIỆN TÍCH 
 
         [StringLength(200)]
         [Column("IMAGE_ANIMAL_STORAGE")]
